Refactored NAnt (build) logging system : moved to an event driver logging system, and allow thresholds to be set.  The build outcome is now completely handled by the loggers.
Replaced the LogListener with an IBuildLogger interface for loggers and an IBuildListener interface for build listeners.

The ConsoleLogger is replaced by a DefaultLogger, which always logs to the console, and which can also be configured to log to a file (using the -logfile option).

The static Log class is removed and replaced by instance methods on the Project and Task classes.  These instance methods actually fire events that get handled by the build logger and buid listeners.

Added a Level enum that is used to assign a priority (level) to a message.  This allows messages to be filtered on their priority.

Added a strongly typed collection for the build listeners.

Updated all existing logging statements to use the new logging system instead of the static Log class.  Assigned levels to the existing logging statements.

Added command-line options for filtering build output (-debug and -quiet).